third_party/blink/web_tests/external/wpt/trusted-types/script-enforcement-004-expected.txt

This is a testharness.js-based test.
[FAIL] Verify when the default policy is called for script set source via Node.textContent.
  assert_array_equals: expected property 1 to be "APPEND" but got "CREATE_SCRIPT" (expected array ["SET", "APPEND", "CREATE_SCRIPT", "HTMLScriptElement text", "RUN_TRUSTED_SCRIPT"] got ["SET", "CREATE_SCRIPT", "HTMLScriptElement textContent", "APPEND", "RUN_TRUSTED_SCRIPT"])
Harness: the test ran to completion.
